Confirmation before delete account

+++ frontend/event-helper/src/pages/UserPage.js
@@ -35,11 +35,13 @@ export default function UserPage(props) {
     }
 
     const deleteAccount = async () => {
-        await fetch(`http://localhost:3000/api/delete-account/${userDetails.sub}`, {
-            method: 'DELETE'
-        })
-        AuthService.logout();
-        window.location.replace("/home")
+        if (window.confirm("Are you sure you want permanently remove your account")) {
+            await fetch(`http://localhost:3000/api/delete-account/${userDetails.sub}`, {
+                method: 'DELETE'
+            })
+            AuthService.logout();
+            window.location.replace("/home")
+        }
     }
 
     return (
